backend/requirements_security.txt
# Security Requirements for Jarvis Command Center V2

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Authentication & Authorization
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Rate Limiting
slowapi==0.1.9

# Security
cryptography==41.0.7
bcrypt==4.1.1

# HTTP Requests
requests==2.31.0
urllib3==2.1.0

# System Monitoring
psutil==5.9.6

# WebSocket
websockets==12.0

# Testing & Security Scanning
bandit==1.7.5
safety==2.3.5
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
